Move ecc compressed marshaling to it's own file
