+++ go/vrf/conversion_test.go
@@ -16,8 +16,6 @@ package vrf
 
 import (
 	"bytes"
-	"crypto/elliptic"
-	"crypto/rand"
 	"fmt"
 	"math/big"
 	"testing"
@@ -48,25 +46,5 @@ func TestI2OSP(t *testing.T) {
 				t.Errorf("I2OSP(%v, %v): %v, want %v", tc.x, tc.xLen, got, tc.want)
 			}
 		})
-	}
-}
-
-func TestCompressedMarshalUnmarshal(t *testing.T) {
-	c := elliptic.P256()
-	_, Ax, Ay, err := elliptic.GenerateKey(c, rand.Reader)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	b := marshalCompressed(c, Ax, Ay)
-	Bx, By, err := unmarshalCompressed(c, b)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if Bx.Cmp(Ax) != 0 {
-		t.Fatalf("Bx: %v, want %v", Bx, Ax)
-	}
-	if By.Cmp(Ay) != 0 {
-		t.Fatalf("By: %v, want %v", By, Ay)
 	}
 }
